backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.25.0
ollama>=0.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
